lambda_folder/requirements.txt
pandas==2.2.3
numpy<2.0
pyarrow==19.0.0
scikit-surprise==1.1.4
beautifulsoup4==4.13.3
lxml==5.3.0
requests==2.32.3
joblib==1.4.2
tqdm==4.67.1